SeedMaker Die README.md
Welcome to SeedMaker Die a Printable 100% Offline Way to Make Your Own Seed Phrase on Paper

It took me week going down the rabbit hole of learning all about how a seed words and i built a whole heap of online spreadsheets that i shared with the public and even used GROK3 to write a SeedMaker Kivy and code etc. At the end of the day i derived this simple but i think the best way for me anyway to know your seed is done properly where you can 100% verify and see that you rolled the die yourself you seen how it was done, you calculated the results you seen how it was all done on open black and white pen and peace of paper.