Create 12 Word Seed
WARNING: This should only be used to have a play. But when you get real and want to create a proper seed in a wallet that you will actually send Bitcoin or alt coins too i strongly suggest using the printable version of SeedMaker Die. 

You;ve read hits before: NEVER PUT A SEED IN A WALLET THAT IS ONLINE ALWAYS USE A HARDWARE WALLET OR SEEDSIGNER DEVICE. 

You can get hacked, ppl get hacked and ppl will be hacked again in the future. If your seed is 100% offline on a hardware device and done properly you minimise your risks. Nothing is ever 100% safe. Humans can make it, humans can break it.