Create 24_Word_Seed.py
Frustrating couple days trying to explain to GROK3 how to build this and it repeatedly screwed it up about 45 times.  But i got it to something usable in the end. 


This is similar format to the offline print page and how it works. Have a play but i would suggest strongly not to use it for any real seed in any online wallet or hardware. 

Use it to learn stuff but do the old way of print and page and roll dice and calculate the seed words and checksum offline 100%. 

The goal with any seed is to do it offline. That is the challenge not as easy as you think. The moment you are on a online pc or even a pc that is offline but will connect to the internet in the future is a bad risk.