Create 24_Word_Seed_Random_Colors.py
Same as others just added a button to generate about 50% of the 0's and changes them to 1s and gives each one a random color. The color has nothing to do with randomness selections it is just as an interesting effect. No 'change colors to black again button' black button like the 12 word variation. Next version variation  will have a few decks of cards to generate randomness and shuffles etc. 

Generally speaking all bits can be either 0 or 1 so it is a 50/50 change that each one will be either a 0 or 1 so its black and white really. If you flip an unbiased coin  for example 128 times and take a tail as 0 and a head as 1 and recorded all them you should have a good random set of 128 bits. This is basically what is happening in any or all these 12 or 24 word seed generation. 256 bits for 24 words of course. 

You can see that JUSt doing this simple exercise on with pen and peace of paper and flipped coin can give you a very valuable life lesson to carry with you in life to do your own seed generation. Do it off a PC or any dam thing that is online and know that it is done right. 

Only use this took online or on your pc to learn about it. Do the REAL seed offline and with pen and paper. Don't be lazy just do it get off that dam computer!!!